Add button press animation and disabled button style
This commit adds a press animation for the choice buttons. When user clicks on their choice, the button will shrink to simulate a press and revert back to original size at end of press.

Additionally, once the game has ended, all buttons will now be greyed out and no hover animation will happen to signal the end of the game.